time parameter now takes in float instead of int

+++ src/Transformers/ArrayTransformer.php
@@ -35,7 +35,7 @@ class ArrayTransformer implements TransformerInterface
     public function transform($data, $limit = 10, $time = 60)
     {
         if (($ip = array_get($data, 'ip')) && ($route = array_get($data, 'route'))) {
-            return new Data((string) $ip, (string) $route, (int) $limit, (int) $time);
+            return new Data((string) $ip, (string) $route, (int) $limit, (float) $time);
         }
 
         throw new InvalidArgumentException('The data array does not provide the required ip and route information.');
